Validity of arguments
The validity of the arguments passed to the program are checked. PID is a
mandatory argument. The core file name is optional, if not specified the
default file name is core.pid.

If a self dump is requested then, the application is called with a --daemon
option or the --socket option.

If the SYSTEMD_SOCKET_OPTION is enabled, then the socket option will be fully
functional. The system where the program is run needs to have sd-daemon.h. For
this we need to have the correct version of systemd which has this socket
option enabled.

Signed-off-by: Janani Venkataraman <jananive@linux.vnet.ibm.com>