Systemd Socket ID retrieval
The socket created by systemd, waits for connections. Once it accepts a
connection, it starts the service which handles the request for a dump.

From this service, we fetch the socket ID and perform the dump.

This is not functional yet, due to the absence of the sd-daemon.h header file.
For this we need to have the correct version of systemd which has this socket
option enabled.

Signed-off-by: Janani Venkataraman <jananive@linux.vnet.ibm.com>