Fetching thread specific Notes
Populates prstatus for each and every thread of the task to be dumped by
accessing their respective /proc/tid/stat.

Fetching the required registers based on the architecture. The architectures
that are handled are X86, S390 and PPC.

We fetch the registers in thread-wise order, as some architectures depend on
the ordering of the notes. (e.g, s390)

Signed-off-by: Suzuki K. Poulose <suzuki@in.ibm.com>
Signed-off-by: Janani Venkataraman <jananive@linux.vnet.ibm.com>